requirements.txt
solana>=0.30.0
solders>=0.16.5
stripe
requests
pyngrok
python-dotenv
uagents